refactor: separate the script movement
